Update GeoClimate version (#1113)
* Update GeoClimate version
Force BBOX coordinates to float

* Update unit test

+++ osm/src/main/groovy/org/orbisgis/geoclimate/osm/OSM.groovy
@@ -71,7 +71,7 @@ abstract class OSM extends AbstractScript {
     static List bbox(Geometry geometry) {
         if (geometry) {
             Envelope env = geometry.getEnvelopeInternal()
-            return [env.getMinY(), env.getMinX(), env.getMaxY(), env.getMaxX()]
+            return [env.getMinY() as float, env.getMinX() as float, env.getMaxY() as float, env.getMaxX() as float]
         }
     }
 
@@ -84,7 +84,7 @@ abstract class OSM extends AbstractScript {
      */
     static List bbox(Envelope env) {
         if (env) {
-            return [env.getMinY(), env.getMinX(), env.getMaxY(), env.getMaxX()]
+            return [env.getMinY() as float, env.getMinX() as float, env.getMaxY() as float, env.getMaxX() as float]
         }
     }
